Points working code
Points are added to the "Game Manager" not the player object. Unsure if this will cause issues in the future, but hey, it works.

Points added when object is picked up, and when enemies are killed.

Added a Serialize Field to the pick up for various points (or point loss)